refactor secret managers
