+++ litellm/router_utils/client_initalization_utils.py
@@ -9,7 +9,7 @@ import openai
 import litellm
 from litellm._logging import verbose_router_logger
 from litellm.llms.azure import get_azure_ad_token_from_oidc
-from litellm.proxy.secret_managers.get_azure_ad_token_provider import (
+from litellm.secret_managers.get_azure_ad_token_provider import (
     get_azure_ad_token_provider,
 )
 from litellm.utils import calculate_max_parallel_requests
