+++ litellm/secret_managers/get_azure_ad_token_provider.py
@@ -14,8 +14,7 @@ def get_azure_ad_token_provider() -> Callable[[], str]:
     Returns:
         Callable that returns a temporary authentication token.
     """
-    from azure.identity import ClientSecretCredential
-    from azure.identity import get_bearer_token_provider
+    from azure.identity import ClientSecretCredential, get_bearer_token_provider
 
     try:
         credential = ClientSecretCredential(
@@ -24,7 +23,9 @@ def get_azure_ad_token_provider() -> Callable[[], str]:
             tenant_id=os.environ["AZURE_TENANT_ID"],
         )
     except KeyError as e:
-        raise ValueError("Missing environment variable required by Azure AD workflow.") from e
+        raise ValueError(
+            "Missing environment variable required by Azure AD workflow."
+        ) from e
 
     return get_bearer_token_provider(
         credential,
