+++ litellm/tests/test_secret_manager.py
@@ -18,7 +18,7 @@ import pytest
 
 from litellm.llms.azure import get_azure_ad_token_from_oidc
 from litellm.llms.bedrock.chat import BedrockConverseLLM, BedrockLLM
-from litellm.proxy.secret_managers.aws_secret_manager import load_aws_secret_manager
+from litellm.secret_managers.aws_secret_manager import load_aws_secret_manager
 from litellm.secret_managers.main import get_secret
